refactor: split lib/types/hooks/game-state.ts into domain-focused type subfiles (#499)
Extract 4 concern groups into gameStateTypes/ subfiles: hooks (UseGameOptionsProps,
UseBaseGameConfig), state (GameState, GameProgress, GameActions, GameResponseActions),
db (GameItem, GameTypeDbRecord), logicState (GameLogicState + its private building
blocks), returns (GameContextHookReturn, UseGameDataReturn). Reduce game-state.ts to
a 6-line barrel re-exporting all public types. No consumer import paths changed.
